style: increase the top and bottom padding of flex container buttons.
style: make all box-shadows more consistent.

refactor: reduce the number of codes the get code button returns, and make it so that font-size, min-width and min-height are only added if modified in the playground.

feat: add a tooltip to the flex item 'flex' property, explaining what each input does.

fix: remove the 'add item' button from the scope of the playground, as it was acting as the first child of the container.